requirements.txt
# Core Framework
langgraph>=0.2.0
langchain>=0.1.0
langchain-community>=0.0.20

# Embeddings (Local, Free)
sentence-transformers>=2.3.0

# Optional: LLM APIs (NOT REQUIRED - only if you want to add LLM features)
# openai>=1.12.0
# anthropic>=0.18.0
# langchain-openai>=0.0.5

# Vector Database
chromadb>=0.4.22
faiss-cpu>=1.7.4

# NER & NLP
spacy>=3.7.0
stanza>=1.7.0

# Database
sqlalchemy>=2.0.25
sqlite3

# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
numpy>=1.26.0
pandas>=2.1.4

# Optional: Sentiment Analysis
vaderSentiment>=3.3.2
textblob>=0.17.1

